+++ js/constants.js
@@ -1,0 +1,34 @@
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Артём', 'Мария', 'Дмитрий', 'Анна', 'Сергей',
+  'Елена', 'Иван', 'Ольга', 'Алексей', 'Наталья',
+  'Павел', 'Юлия', 'Михаил', 'Екатерина', 'Андрей'
+];
+
+const DESCRIPTIONS = [
+  'Прекрасный закат на море',
+  'Горный пейзаж в утреннем тумане',
+  'Улочки старого города',
+  'Кофе в уютной кофейне',
+  'Прогулка по осеннему лесу',
+  'Архитектура современного мегаполиса',
+  'Мой питомец в смешной позе',
+  'Вкусный ужин в ресторане',
+  'Путешествие по неизведанным местам',
+  'Творческий процесс на работе',
+  'Отдых на пляже в летний день',
+  'Зимняя сказка в парке',
+  'Цветущий сад весной',
+  'Ночной город с высоты',
+  'Спортивные достижения'
+];
+
+export { MESSAGES, NAMES, DESCRIPTIONS };
