+++ js/data.js
@@ -1,0 +1,28 @@
+import { MESSAGES, NAMES, DESCRIPTIONS } from './constants.js';
+import { getRandomInteger, getRandomArrayElement, createIdGenerator } from './util.js';
+
+const generateCommentId = createIdGenerator();
+
+const createComment = () => ({
+  id: generateCommentId(),
+  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
+  message: getRandomArrayElement(MESSAGES),
+  name: getRandomArrayElement(NAMES)
+});
+
+const createComments = () => {
+  const commentsCount = getRandomInteger(0, 30);
+  return Array.from({ length: commentsCount }, createComment);
+};
+
+const createPhoto = (index) => ({
+  id: index,
+  url: `photos/${index}.jpg`,
+  description: getRandomArrayElement(DESCRIPTIONS),
+  likes: getRandomInteger(15, 200),
+  comments: createComments()
+});
+
+const generatePhotos = () => Array.from({ length: 25 }, (_, index) => createPhoto(index + 1));
+
+export { generatePhotos };
