+++ js/main.js
@@ -1,74 +1,4 @@
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAMES = [
-  'Артём', 'Мария', 'Дмитрий', 'Анна', 'Сергей',
-  'Елена', 'Иван', 'Ольга', 'Алексей', 'Наталья',
-  'Павел', 'Юлия', 'Михаил', 'Екатерина', 'Андрей'
-];
-
-const DESCRIPTIONS = [
-  'Прекрасный закат на море',
-  'Горный пейзаж в утреннем тумане',
-  'Улочки старого города',
-  'Кофе в уютной кофейне',
-  'Прогулка по осеннему лесу',
-  'Архитектура современного мегаполиса',
-  'Мой питомец в смешной позе',
-  'Вкусный ужин в ресторане',
-  'Путешествие по неизведанным местам',
-  'Творческий процесс на работе',
-  'Отдых на пляже в летний день',
-  'Зимняя сказка в парке',
-  'Цветущий сад весной',
-  'Ночной город с высоты',
-  'Спортивные достижения'
-];
-
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
-const createIdGenerator = () => {
-  let lastGeneratedId = 0;
-  return () => {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-};
-
-const generateCommentId = createIdGenerator();
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: getRandomArrayElement(MESSAGES),
-  name: getRandomArrayElement(NAMES)
-});
-
-const createComments = () => {
-  const commentsCount = getRandomInteger(0, 30);
-  return Array.from({ length: commentsCount }, createComment);
-};
-
-const createPhoto = (index) => ({
-  id: index,
-  url: `photos/${index}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(15, 200),
-  comments: createComments()
-});
-
-const generatePhotos = () => Array.from({ length: 25 }, (_, index) => createPhoto(index + 1));
+import { generatePhotos } from './photo-generator.js';
+import { getRandomInteger, getRandomArrayElement } from './util.js';
 
 export { generatePhotos, getRandomInteger, getRandomArrayElement };
